05.12.2020

+++ src/main/java/uni/fmi/masters/controller/CommentController.java
@@ -7,6 +7,7 @@ import javax.servlet.http.HttpSession;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.annotation.Secured;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -67,7 +68,9 @@ public class CommentController {
 		return commentRepo.findAll();
 	}
 	
+	
 	@DeleteMapping(path = "/comment/delete")
+	@Secured("ROLE_ADMIN")
 	public ResponseEntity<Boolean> deleteComment(
 			@RequestParam(value = "id")int id,
 			HttpSession session
@@ -84,9 +87,13 @@ public class CommentController {
 		if(optionalComment.isPresent()) {
 			CommentBean comment = optionalComment.get();
 			
-			commentRepo.delete(comment);
-			
-			return new ResponseEntity<>(true, HttpStatus.OK);
+			if(comment.getUser().getId() == user.getId())	{
+				commentRepo.delete(comment);				
+				return new ResponseEntity<>(true, HttpStatus.OK);
+			}else {
+				return new ResponseEntity<>(false, HttpStatus.I_AM_A_TEAPOT);
+			}
+
 			
 			
 		}else {
